mitosheet: remove unecessary code
