+++ mitosheet/src/jupyter/api.tsx
@@ -6,7 +6,8 @@ import { GraphObject } from "../components/taskpanes/ControlPanel/SummaryStatsTa
 import { UniqueValueCount, UniqueValueSortType } from "../components/taskpanes/ControlPanel/ValuesTab/ValuesTab";
 import { FileElement } from "../components/taskpanes/Import/ImportTaskpane";
 import { valuesArrayToRecord } from "../components/taskpanes/PivotTable/pivotUtils";
-import { BackendPivotParams, FrontendPivotParams, SplitTextToColumnsParams } from "../types";
+import { SplitTextToColumnsParams } from "../components/taskpanes/SplitTextToColumns/SplitTextToColumnsTaskpane";
+import { BackendPivotParams, FrontendPivotParams } from "../types";
 import { ColumnID, ExcelFileMetadata, FeedbackID, FilterGroupType, FilterType, FormatTypeObj, GraphID, MitoError, SearchMatches, GraphParams } from "../types";
 import { getDeduplicatedArray } from "../utils/arrays";
