+++ mitosheet/src/types.tsx
@@ -270,15 +270,6 @@ export interface ConcatParams {
     sheet_indexes: number[]
 }
 
-export interface SplitTextToColumnsParams {
-    sheet_index: number,
-    column_id: ColumnID | undefined,
-    delimiters: string[],
-    // Note: We create the new_column_header_suffix on the frontend so that it is saved in the step parameters, 
-    // which allows us to replay the analysis and generate the same columns. 
-    new_column_header_suffix: string 
-}
-
 // NOTE: these aggregation functions need to be supported
 // in mitosheet/steps/pivot.py as well
 export enum AggregationType {
